# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Healthcare Data Mapping Tool API",
    "description": "\nv1 REST API for the Healthcare Data Mapping Tool.\n\n## Auth (v1 simulated RBAC)\nPass `X-User-Id` and `X-User-Role` headers with every request.\nRoles: `viewer`, `analyst`, `approver`, `admin`.\n\n## Phase scope\n- Schema import + catalog (Phase 3)\n- Mapping suggestions (Phase 4)\n- Mapping validation with configurable policy (Phase 4)\n- Mapping artifact export (Phase 6 script placeholder)\n\n## Out of scope for v1\n- Transformation execution (Phase 6)\n- Patient-level data ingestion (forbidden in v1)\n",
    "version": "1.0.0"
  },
  "paths": {
    "/projects": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "operationId": "create_project_projects_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_projects_get",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response List Projects Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project Status",
        "operationId": "update_project_status_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/schemas/source/import": {
      "post": {
        "tags": [
          "schemas"
        ],
        "summary": "Import Source Schema",
        "description": "Import a source schema via the adapter registry, register in catalog, pin to project.",
        "operationId": "import_source_schema_projects__project_id__schemas_source_import_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSourceSchemaRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinnedSchemaRef"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/schemas/target/select": {
      "post": {
        "tags": [
          "schemas"
        ],
        "summary": "Select Target Schema",
        "description": "Pin an already-registered catalog schema as the target for this project.",
        "operationId": "select_target_schema_projects__project_id__schemas_target_select_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectTargetSchemaRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinnedSchemaRef"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/schemas": {
      "get": {
        "tags": [
          "schemas"
        ],
        "summary": "Get Project Schemas",
        "description": "Return pinned source and target schema references for a project.",
        "operationId": "get_project_schemas_projects__project_id__schemas_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "anonymous",
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "viewer",
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSchemasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sensitivity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensitivity",
            "default": "healthcare_highly_sensitive"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateProjectRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportSourceSchemaRequest": {
        "properties": {
          "format_id": {
            "type": "string",
            "title": "Format Id",
            "description": "Adapter format identifier. Examples: 'json_schema', 'csv_schema'. Open string — no hardcoded list."
          },
          "schema_name": {
            "type": "string",
            "minLength": 1,
            "title": "Schema Name"
          },
          "version_label": {
            "type": "string",
            "minLength": 1,
            "title": "Version Label"
          },
          "schema_payload": {
            "title": "Schema Payload",
            "description": "Raw schema payload to parse. Shape depends on format_id: JSON object for json_schema, CSV string or list-of-dicts for csv_schema."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "format_id",
          "schema_name",
          "version_label",
          "schema_payload"
        ],
        "title": "ImportSourceSchemaRequest"
      },
...
      "PinnedSchemaRef": {
        "properties": {
          "pin_id": {
            "type": "string",
            "title": "Pin Id"
          },
          "schema_id": {
            "type": "string",
            "title": "Schema Id"
          },
          "version_label": {
            "type": "string",
            "title": "Version Label"
          },
          "schema_name": {
            "type": "string",
            "title": "Schema Name"
          },
          "format_id": {
            "type": "string",
            "title": "Format Id"
          },
          "field_count": {
            "type": "integer",
            "title": "Field Count"
          }
        },
        "type": "object",
        "required": [
          "pin_id",
          "schema_id",
          "version_label",
          "schema_name",
          "format_id",
          "field_count"
        ],
        "title": "PinnedSchemaRef"
      },
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "sensitivity": {
            "type": "string",
            "title": "Sensitivity"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "status",
          "owner",
          "sensitivity",
          "created_at"
        ],
        "title": "ProjectResponse"
      },
      "ProjectSchemasResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PinnedSchemaRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PinnedSchemaRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "source",
          "target"
        ],
        "title": "ProjectSchemasResponse"
      },
      "SelectTargetSchemaRequest": {
        "properties": {
          "schema_id": {
            "type": "string",
            "title": "Schema Id",
            "description": "Catalog schema_id of the schema to pin as target."
          },
          "version_label": {
            "type": "string",
            "minLength": 1,
            "title": "Version Label"
          }
        },
        "type": "object",
        "required": [
          "schema_id",
          "version_label"
        ],
        "title": "SelectTargetSchemaRequest"
      },
...
      "UpdateProjectStatusRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Target status. Allowed: review, approved, published, archived"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "UpdateProjectStatusRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}